backend/requirements.txt
asgiref==3.4.1
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Django==3.2.17
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.10.0
pyparsing==3.0.9
pytz==2022.7.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.6.0
